fix counter
